bee-storage: move tests and impl under sqlx dir
